requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
DBUtils==3.0.3
idna==3.6
mysql-connector-python==8.3.0
requests==2.31.0
tenacity==8.2.3
urllib3==2.1.0
